vim-patch:7.4.1986
Problem:    Compiler warns for loss of data.
Solution:   Use size_t instead of int. (Christian Brabandt)

https://github.com/vim/vim/commit/fef524bbff9aa186838c35212b2f89f61d627cf8

+++ src/nvim/ex_cmds2.c
@@ -2485,7 +2485,7 @@ static void add_pack_plugin(char_u *fname, void *cookie)
     *p4 = NUL;
 
     // Find "ffname" in "p_rtp", ignoring '/' vs '\' differences
-    int fname_len = STRLEN(ffname);
+    size_t fname_len = STRLEN(ffname);
     char_u *insp = p_rtp;
     for (;;) {
       if (vim_fnamencmp(insp, ffname, fname_len) == 0) {
